- Updated a lot of stuffs, and i don't remember
What I can remember:
- Added InputHandler & MouseHandler
- MouseHover feature is setup
- Menu Buttons is bit more flexible now

+++ SDL Class Project/Vector.cpp
@@ -6,7 +6,7 @@ Vector::Vector() {
 
 }
 
-Vector::Vector(float _x, float _y) {
+Vector::Vector(int _x, int _y) {
 	this->x = _x;
 	this->y = _y;
